Protectionism (Wealth) policy nerfed
Removed the extra 10% gold and production from the Protectionism policy.
Retains +1 production for mines and lumbermills in addition to +1 gold
for villages.  This policy is very strong in that the extra gold and
production on the improvements is multiplied by other factors, such as
buildings like the Market, Bank, Smith, Workshop, Factory, etc.  I've
been playing with this setting for a while and it's quite balanced with
the other policies in the tree.