Fixing poor Google translation
Google Translate falls over badly with the variables defined in curly
braces. Especially in the Western European languages. The Slavic
languages seem to handle it quite well.
The Asian languages are going to be tricky.
About the only errors I can pick out is in Chinese the [NEWLINE]
variable gets translated.