Representation Update #2
Upon doing further testing on this policy in Liberty I discovered what
it's doing.  The vanilla code grants +1 happiness per connected city.
The mod code grants an additional +1 happiness per connected city in
addition to +1 happiness in the capital.  Not sure if this is the
intended effect, but that it what it's doing right now.  This updates
the text to reflect what the code is doing.