Representation update #3 & Meritocracy Update
The effects of these policies are overlapping with each other.  Both
policies affect culture costs per city and happiness per city
connection.  Finally got to the bottom of what's going on.  This commits
updates the text to accurately state what they are doing.  Pretty sure
it's an unintended bug.  I can fix it if you want me to.