Update CEF_Data.xml
The Yield_Change to Culture of 16 here is replaced by a change in lines 253 -- 259 where it is set to 10.
The vanilla value is 5 so I am assuming this is the more reasonable.
Why the change is made that way by DELETING it first I don't know. There was always only one reference to this BELIEF in this table and it was always to the Hermitage.
Then we had the same code block commented out immediately below it?! I removed that, seems overly pointless to keep all the changes in the file. GitHub keeps track of all this.